requirements.txt
asgiref==3.4.1
astroid==2.8.3
black==21.9b0
click==8.0.3
Django==3.2.5
django-debug-toolbar==3.2.2
django-extensions==3.1.3
django-rest-framework==0.1.0
django-typeahead==1.0.1
djangorestframework==3.12.4
isort==5.9.3
jsonpickle==2.0.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
pylint==2.11.1
pytz==2021.1
regex==2021.10.23
sqlparse==0.4.1
toml==0.10.2
tomli==1.2.2
typing-extensions==3.10.0.2
wrapt==1.13.2
